requirements.txt
alembic==1.16.4
fastapi==0.116.1
requests==2.31.0
psycopg2==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.41
uvicorn==0.35.0
